refactor: rename to traefik_route_types
